+++ sync-for-payables/src/codat_sync_for_payables/sdkconfiguration.py
@@ -27,9 +27,9 @@ class SDKConfiguration:
     server_idx: Optional[int] = 0
     language: str = "python"
     openapi_doc_version: str = "3.0.0"
-    sdk_version: str = "7.0.0"
-    gen_version: str = "2.445.1"
-    user_agent: str = "speakeasy-sdk/python 7.0.0 2.445.1 3.0.0 codat-sync-for-payables"
+    sdk_version: str = "8.0.0"
+    gen_version: str = "2.463.0"
+    user_agent: str = "speakeasy-sdk/python 8.0.0 2.463.0 3.0.0 codat-sync-for-payables"
     retry_config: OptionalNullable[RetryConfig] = Field(default_factory=lambda: UNSET)
     timeout_ms: Optional[int] = None